% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coordinates_from_fasta.R
\name{extract_coordinates_from_fasta}
\alias{extract_coordinates_from_fasta}
\title{Extract coordinates from fasta files}
\usage{
extract_coordinates_from_fasta(fasta_file)
}
\arguments{
\item{fasta_file}{fasta files.}
}
\value{
coordinates_df
}
\description{
Extract coordinates from fasta files
}
\examples{
\dontrun{extract_coordinates_from_fasta(fasta_file)}
}
