% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_manager.R
\name{package_manager}
\alias{package_manager}
\title{Install and load packages}
\usage{
package_manager(packages)
}
\arguments{
\item{packages}{vector of packages to load.}
}
\value{
Nothing.
}
\description{
Install and load packages
}
\examples{
package_manager(c("shiny"))
}
