% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMultiFasta.R
\name{readMultiFasta}
\alias{readMultiFasta}
\title{Read fasta files}
\usage{
readMultiFasta(file_path)
}
\arguments{
\item{file_path}{path to the file.}
}
\value{
sequences
}
\description{
Read fasta files
}
\examples{
\dontrun{sequences <- readMultiFasta("path/to/your/file.fasta")}


}
